requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
duckduckgo_search==5.3.1b1
greenlet==3.1.0
gTTS==2.5.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jaraco.context==6.0.1
more-itertools==10.5.0
openai==1.29.0
pillow==10.3.0
psycopg-pool==3.2.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytube==15.0.0
redis==5.0.8
regex==2024.9.11
requests==2.31.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
spotipy==2.23.0
SQLAlchemy==2.0.35
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
whois==0.9.27
wolframalpha==5.0.1
xmltodict==0.13.0
